requirements.txt
dnspython
PyYAML
requests